Backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
pymysql==1.1.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt,argon2]==1.7.4
fastapi-mail==1.4.1
jinja2==3.1.4

supabase==2.4.0
